Fixed a ClassNotFoundException causing the application startup to fail when a (unused) plugin category class could not be loaded.
